Collection of helper scripts from diagnosing the "slow BIND" problem. Pathology is triggered by running many concurrent simple queries involving the same joins using either simple statements or unnamed prepared statements with one or more bind variables, so that each run of the query requires evaluating execution plans.
Capturing and aggregating many stack traces shows the dominant paths in the postgres code stack
where the time is spent.

This highlights that merge join selectivity estimation tends to always trigger dynamic checking
of the upper histogram bounds against actual max value.  (Presumably this is to compensate for
possibly stale statistics.)  This call to BufferAlloc tends to stall trying to get the LWLock
in SHARE mode on one of the 16 partitions of the buffer cache.