Add new pgbench-based test case using tables with histograms on non-uniform distribution of join-key values.
The standard 4 pgbench tables lack histograms and indexes (and constraints) on their foreign key columns.
To emulate our real-world application behavior, we need to reach the code path to "get_actual_variable_range",
as indirectly invoked by "mergejoinscansel" through "scalarineqsel".  However, during "mergejoinscansel" for
a query joining two of the standard pgbench tables, the lack of both a histogram and MCV on either of the
join keys causes us to skip the call to "scalarineqsel".  This is of course appropriate, since there is
nothing to do in "scalarineqsel" and its descendents without histogram bounds to compare.

So we must setup a pair of tables with a non-uniform distribution of values in the join key column, so that
the ANALYZE deems it worthwhile to keep histogram (and MCV) statistics on these columns.